pyproject.toml
[tool.poetry]
name = "getstockprices"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
prices = "^1.1.0"
flask = "^1.1.2"
discord = "^1.0.1"
timeloop = "^1.0.2"
schedule = "^1.1.0"
pandas = "^1.2.4"
tabulate = "^0.8.9"
pymongo = "^3.11.4"
dnspython = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
